requirements.txt
scapy
sqlalchemy
psycopg2-binary
fastapi
uvicorn
pydantic
dearpygui
python-dotenv
requests